backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.3
requests==2.31.0
stripe==6.7.0
twilio==8.10.0
typeguard==4.4.4
pydantic==2.11.7